requirements.txt
opencv-python-headless
mediapipe
pyautogui
numpy
pywebview